Remove redundant show-toast example component
