+++ apps/cookbook/src/app/examples/_modal-playground.shared.scss
@@ -2,7 +2,6 @@
   display: grid;
   place-content: center;
   overflow-y: auto;
-  height: 100%;
 
   > button {
     min-width: 15rem;
