+++ apps/cookbook/src/app/examples/examples.routes.ts
@@ -89,7 +89,6 @@ import { ModalExampleAlertWithGuardComponent } from './modal-example/modal-examp
 import { ModalEmbeddedAlertExampleComponent } from './modal-example/alert-example/modal-example-embedded-alert.component';
 import { ModalComponentExampleComponent } from './modal-example/modal-component-example.component';
 import { SlidesHeightExampleComponent } from './slides-example/slides-height-example/slides-height-example.component';
-import { ShowToastExampleComponent } from './toast-example/examples/show-toast-example';
 
 export const EXAMPLE_ROUTES: ModalEnabledRoutes = [
   {
@@ -351,12 +350,6 @@ export const EXAMPLE_ROUTES: ModalEnabledRoutes = [
       {
         path: 'toast',
         component: ToastExampleComponent,
-        children: [
-          {
-            path: 'show-toast',
-            component: ShowToastExampleComponent,
-          },
-        ],
       },
     ],
   },
