+++ apps/cookbook/src/app/showcase/toast-showcase/toast-showcase.component.html
@@ -12,14 +12,14 @@
   </ul>
 
   <cookbook-iphone
-    src="/examples/toast/show-toast"
+    src="/examples/toast"
     showExternalLink="true"
     viewMode="full-size"
     showViewModeToggle="true"
   ></cookbook-iphone>
 
   <!-- prettier-ignore -->
-  <cookbook-example-viewer [ts]="showToastExampleComponent.ts"></cookbook-example-viewer>
+  <cookbook-example-viewer [ts]="toastExample.ts"></cookbook-example-viewer>
   <h4>Callback</h4>
   If you need to know when the toast is dismissed, you can pass a callback function:
   <!-- prettier-ignore -->
