+++ apps/cookbook/src/app/showcase/toast-showcase/toast-showcase.component.ts
@@ -1,12 +1,10 @@
 import { Component } from '@angular/core';
-import exampleHtml from '../../examples/toast-example/toast-example.component.html?raw';
-import { ToastExampleComponent } from '../../examples/toast-example/toast-example.component';
 import { CodeViewerComponent } from '../../shared/code-viewer/code-viewer.component';
 import { ApiDescriptionPropertiesComponent } from '../../shared/api-description/api-description-properties/api-description-properties.component';
 import { ExampleViewerComponent } from '~/app/shared/example-viewer/example-viewer.component';
 import { ApiDescriptionProperty } from '~/app/shared/api-description/api-description-properties/api-description-properties.component';
 import { IphoneComponent } from '~/app/iphone/iphone.component';
-import { ShowToastExampleComponent } from '~/app/examples/toast-example/examples/show-toast-example';
+import { ToastExampleComponent } from '~/app/examples/toast-example/toast-example.component';
 
 @Component({
   selector: 'cookbook-toast-showcase',
@@ -21,8 +19,8 @@ import { ShowToastExampleComponent } from '~/app/examples/toast-example/examples
   ],
 })
 export class ToastShowcaseComponent {
-  exampleHtml = exampleHtml;
-  showToastExampleComponent = ShowToastExampleComponent;
+  toastExample = ToastExampleComponent;
+
   properties: ApiDescriptionProperty[] = [
     {
       name: 'message',
